add plans module
