+++ app/(root)/(tabs)/meals.tsx
@@ -1,7 +1,7 @@
 import { Heading } from "@/components/ui/heading";
 import { VStack } from "@/components/ui/vstack";
 
-export default function Files() {
+export default function Meals() {
   return (
     <VStack>
       <Heading className="text-typography-900">Coming soon</Heading>
