--- app/(root)/(tabs)/plans.tsx
@@ -1,9 +1,0 @@
-import { Heading } from "@/components/ui/heading";
-import { VStack } from "@/components/ui/vstack";
-export default function Plans() {
-  return (
-    <VStack>
-      <Heading className="text-typography-900">Coming soon</Heading>
-    </VStack>
-  );
-}
