+++ app/(root)/(tabs)/plans/community.tsx
@@ -1,0 +1,9 @@
+import { Heading } from "@/components/ui/heading";
+import { VStack } from "@/components/ui/vstack";
+export default function CommunityPlans() {
+    return (
+        <VStack>
+            <Heading className="text-typography-900">Community plans</Heading>
+    </VStack>
+    );
+}
