+++ app/(root)/(tabs)/plans/company.tsx
@@ -1,0 +1,9 @@
+import { Heading } from "@/components/ui/heading";
+import { VStack } from "@/components/ui/vstack";
+export default function CompanyPlans() {
+    return (
+        <VStack>
+            <Heading className="text-typography-900">Coming soon</Heading>
+        </VStack>
+    );
+}
