+++ app/(root)/(tabs)/scanner.tsx
@@ -1,6 +1,6 @@
 import { Heading } from "@/components/ui/heading";
 import { VStack } from "@/components/ui/vstack";
-export default function Profile() {
+export default function Scanner() {
   return (
     <VStack>
       <Heading className="text-typography-900">Coming soon</Heading>
