+++ utils/enum/user-details.enum.ts
@@ -1,0 +1,17 @@
+export enum WeightUnitEnum {
+    KG = 'KG',
+    LBS = 'LBS',
+    ST = 'ST',
+}
+
+export enum HeightUnitEnum {
+    CM = 'CM',
+    IN = 'IN',
+    FT = 'FT',
+}
+
+export enum GoalEnum {
+    WEIGHT_LOSS = 'WEIGHT_LOSS',
+    MAINTAIN = 'MAINTAIN',
+    GAIN_MUSCLE = 'GAIN_MUSCLE',
+}
